addon rename & ammo framework
