+++ addons/orbis_artillery_config/XEH_postInit.sqf
