+++ addons/orbis_artillery_config/XEH_preInit.sqf
