+++ addons/orbis_artillery_config/scripts/define_artillery.hpp
