+++ addons/orbis_artillery_config/scripts/fnc_rangeTablePreCalculatedValues.sqf
