+++ addons/orbis_artillery_config/scripts/script_component.hpp
@@ -1,0 +1,2 @@
+#include "\z\ace\addons\main\script_mod.hpp"
+#include "\z\ace\addons\main\script_macros.hpp"
