--- addons/orbis_rhs_artillery_config/config.cpp
@@ -1,5 +1,0 @@
-#include "CfgPatches.hpp"
-
-#include "CfgEventHandlers.hpp"
-
-#include "CfgVehicles.hpp"
