--- addons/orbis_rhs_artillery_config/script_component.hpp
@@ -1,2 +1,0 @@
-#include "\z\ace\addons\main\script_mod.hpp"
-#include "\z\ace\addons\main\script_macros.hpp"
